very rough concept of multiplayer game
I made a draft. a lot of improvements are needed. now AImove method doesn't work all the time but thats not the point. meybe we will replace it by the smart algorithm now its completely random. exept that it requires other changes inside that mode. we can also improve design of the startingscreeen